+++ test/route/mount_test.exs
@@ -8,13 +8,23 @@ defmodule Maru.Route.MountTest do
       end
     end
 
+    defmodule MountedAlias do
+      def __routes__ do
+        [%Maru.Route{}]
+      end
+    end
+
     defmodule MountTest do
       use Maru.Router
+      alias Maru.Route.MountTest.MountedAlias
+
       mount Maru.Route.MountTest.Mounted
+      mount MountedAlias
 
       def m, do: @mounted
     end
 
-    assert [%Maru.Route{}] = MountTest.m
+    assert [%Maru.Route{}, %Maru.Route{}] = MountTest.m
   end
+
 end